UI: Add context provider to access app state globally
